Refactor TeamCity build to create two subprojects, for the admin app web and admin app installer
