+++ .teamcity/Installers/Project.kt
@@ -1,0 +1,28 @@
+// SPDX-License-Identifier: Apache-2.0
+// Licensed to the Ed-Fi Alliance under one or more agreements.
+// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+// See the LICENSE and NOTICES files in the project root for more information.
+
+package lib
+
+import jetbrains.buildServer.configs.kotlin.v2019_2.*
+
+object AdminAppInstallerProject : Project({
+    name = "Admin App Installer"
+    id = RelativeId("AdminAppInstaller")
+    description = "ODS Admin App Installer Build Configurations"
+
+    params {
+        param("build.feature.freeDiskSpace", "2gb")
+        param("git.branch.default", "main")
+        param("git.branch.specification", """
+            +:refs/heads/(*)
+            +:refs/(pull/*)/merge
+        """.trimIndent())
+        param("teamcity.ui.settings.readOnly","true")
+        param("adminApp.version", "2.2.0")
+    }
+
+    buildType(_self.buildTypes.BuildAdminAppInstaller)
+    buildType(_self.buildTypes.Deploy)
+})
