+++ .teamcity/_Self/Project.kt
@@ -8,25 +8,8 @@ package _self
 import jetbrains.buildServer.configs.kotlin.v2019_2.*
 
 object AdminAppProject : Project({
-    description = "ODS Admin App Build Configurations"
+    description = "Build configurations managed through the Admin App repository"
 
-    params {
-        param("build.feature.freeDiskSpace", "2gb")
-        param("git.branch.default", "main")
-        param("git.branch.specification", """
-            +:refs/heads/(*)
-            +:refs/(pull/*)/merge
-        """.trimIndent())
-        param("teamcity.ui.settings.readOnly","true")
-        param("adminApp.version", "2.2.0")
-    }
-
-    template(_self.templates.BuildAndTestTemplate)
-
-    buildType(_self.buildTypes.BuildBranch)
-    buildType(_self.buildTypes.BuildPullRequests)
-    buildType(_self.buildTypes.Deploy)
-    buildType(_self.buildTypes.PublishPackagesToAzureArtifacts)
-    buildType(_self.buildTypes.PublishReleasePackagesToAzureArtifacts)
-    buildType(_self.buildTypes.BuildAdminAppInstaller)
+    subProject(web.AdminAppWebProject)
+    subProject(installers.AdminAppInstallerProject)
 })
